mysite/staticfiles/static/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
google-auth==2.27.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.6
